Fixed important logical bug in the survsurv sensitivity analysis
The issue was that the wrong knots of the Royston-Parmar model were passed to the quantile function for S_0. This caused weird behaviour for the survival classification proportions. More importantly, the main results of the sensitivity analysis were also wrong as a consequence. This has now been fixed.

The corresponding tests should still be adapted.