Fix issue with sampling rotation parameters in the sensitivity analysis
The issue came up with the Clayton and Gumbel copulas. For these copulas, when the range of Spearman rho parameters was strictly positive in sensitivity analysis, the rotation parameters were still sampled. This meant that negative associations were still possible. This has now been fixed by allowing the user to specify his/her own rotation parameters to be sampled in the sensitivity analysis.